+++ games/2dRacer/AdvancedDemo/Resources/animations/roadLineAnimation.txt
