+++ games/2dRacer/AdvancedDemo/Resources/images/road_mk_dotted.png
