+++ games/2dRacer/AdvancedDemo/Resources/json/Controls.json
@@ -1,0 +1,31 @@
+{
+    "Controls" :
+    ["left", "right", "up", "down", "btn1", "btn2", "btn3", "btn4", "btn5", "btn6"],
+    "Controls_Schemes" : 
+    [
+        {
+            "left"  : "Left",
+            "right" : "Right",
+            "up"    : "Up",
+            "down"  : "Down",
+            "btn1"  : "1",
+            "btn2"  : "2",
+            "btn3"  : "3",
+            "btn4"  : "4",
+            "btn5"  : "5",
+            "btn6"  : "6"
+        },
+        {
+            "left"  : "A",
+            "right" : "D",
+            "up"    : "W",
+            "down"  : "S",
+            "btn1"  : "1",
+            "btn2"  : "2",
+            "btn3"  : "3",
+            "btn4"  : "4",
+            "btn5"  : "5",
+            "btn6"  : "6"
+        }
+    ]
+}
